Move version to standalone file
Before this change in order to bump the operator version one would need
to edit `Makefile` and `manifests/operator.yaml` to set the operator
version. Also the OpenShift metadata bundle version has to be set and
someone has to run a command to generate a new directory with metadata
files.

This patch moves the version into a separate file to make it more
automation friendly. At the same time it starts using the same version
(except the "v" prefix and the "-rcN" suffix) for OpenShift metadata
bundle.

Added a make target that automates a regular release preparation step.